optionally install xp-framework/core as global Composer package
This adds the possibility to install xp-framework/core as global
Composer package when installing xp-runners. Without this package
installation is not complete, and a user must do this step manually.

However, it requires that Composer is already installed, most
likely through another Puppet module. Installing Composer can and
should not be part of this module.

In order to fulfil the task two new parameters are introduced:
- composer_home is required in order to run Composer correctly, as
  Composer otherwise doesn't know which path to use to install the
  package into. (On regular command line it uses the $HOME or
  $COMPOSER_HOME environment variables, but we don't have this when
  running Composer via Puppet exec.)
- composer_path should be set if the Composer binary is not in the
  regular location.

Things to still think about:
- Probably the exec should not run when xp-framework/core is already
  installed globally.
- When we tackle issue #2 (later in another branch) it should be
  possible to define the version of xp-framework/core as well.